# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Personal Library API",
    "description": "API documentation for the Personal Library project"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
    "/users": {
      "post": {
        "summary": "Create a new user",
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "description": "User object that needs to be added",
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "type": "string",
                  "example": "John"
                },
                "lastName": {
                  "type": "string",
                  "example": "Doe"
                }
              },
              "required": ["firstName", "lastName"]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User created successfully"
          }
        }
      },
      "get": {
        "summary": "Retrieve all users",
        "responses": {
          "200": {
            "description": "A list of users"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Retrieve a specific user by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "description": "ID of the user to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved successfully"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "put": {
        "summary": "Update a user by ID",
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "description": "ID of the user to update"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "description": "Updated user object",
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "type": "string",
                  "example": "Jane"
                },
                "lastName": {
                  "type": "string",
                  "example": "Smith"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User updated successfully"
          }
        }
      },
      "delete": {
        "summary": "Delete a user by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "description": "ID of the user to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully"
          }
        }
      }
    },
    "/books": {
      "post": {
        "summary": "Create a new book",
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "description": "Book object that needs to be added",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "The Great Gatsby"
                },
                "author": {
                  "type": "string",
                  "example": "F. Scott Fitzgerald"
                },
                "isbn": {
                  "type": "string",
                  "example": "9780743273565"
                },
                "category": {
                  "type": "string",
                  "example": "Fiction"
                }
              },
              "required": ["title", "author", "isbn"]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Book created successfully"
          }
        }
      },
      "get": {
        "summary": "Retrieve all books",
        "responses": {
          "200": {
            "description": "A list of books"
          }
        }
      }
    },
    "/books/{id}": {
      "get": {
        "summary": "Retrieve a specific book by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "description": "ID of the book to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "Book retrieved successfully"
          },
          "404": {
            "description": "Book not found"
          }
        }
      },
      "put": {
        "summary": "Update a book by ID",
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "description": "ID of the book to update"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "description": "Updated book object",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "Updated Title"
                },
                "author": {
                  "type": "string",
                  "example": "Updated Author"
                },
                "isbn": {
                  "type": "string",
                  "example": "1234567890"
                },
                "category": {
                  "type": "string",
                  "example": "Updated Category"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Book updated successfully"
          }
        }
      },
      "delete": {
        "summary": "Delete a book by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "description": "ID of the book to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Book deleted successfully"
          }
        }
      }
    }
  }
}